style: 전역 변수 값 enum으로 이동
